pyproject.toml
[project]
name = "pixeljumble"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.13.5",
    "discord-py-self",
    "python-dotenv>=1.2.2",
    "torch>=2.11.0",
    "torchvision>=0.26.0",
    "tqdm>=4.66.5",
]

[[tool.uv.index]]
url = "https://download.pytorch.org/whl/cpu"

[tool.uv.sources]
discord-py-self = { git = "https://github.com/dolfies/discord.py-self.git" }
